fixing clock_gettime

+++ perf.h
@@ -1,3 +1,3 @@
 void perf_init();
 
-void flop_compute(char* message, unsigned long long int nb_operation, unsigned long long int cycles);
+void flop_compute(char* message, unsigned long long int nb_operation, double cycles);
